:heavy_plus_sign: Move AP entering into dial function
The dial function will now check to see if the outcome cell is occupied 
if it is then it moves down to the next row and dials. Secondly if the 
column to the left (date of call) is filled then it will assume it was 
just a AP and fill out accordingly.